// app/services/product/product.interface.ts
export interface IProduct {
  id: number
  createdAt: string
  updatedAt: string
  slug: string
  manufacturer: string
  capacity: string
  shape: string
  collection: string
  transformationMechanism: string
  hardnessLevel: string
  upholsteryMaterial: string
  frameMaterial: string
  legMaterial: string
  color: string
  filler: string
  legColor: string
  size: string
  cornerLocation: string
  description: string
  design: string
  springBlock: string
  caseMaterial: string
  price: number
  articul: number
  title: string
  priceWithDiscount: number
  productPath: string
}